PF-26 implement expand of side bar navigation
