+++ src/common/action-button/styles.scss
@@ -7,12 +7,11 @@
   cursor: pointer;
   border-radius: 4px;
 
-  font-style: normal;
-  font: $font-body-bold;
-
   display: flex;
   align-items: center;
   justify-content: center;
+  font-style: normal;
+  font: $font-body-bold;
 
   svg {
     margin-right: 8px;
