+++ src/common/navigation/calendar/date-header/index.jsx
@@ -1,0 +1,31 @@
+import React from "react";
+
+import { IconButton } from "common/icon-button";
+
+import { currYearName, currMonthName } from "../helpers";
+
+import "./styles.scss";
+
+export const DateHeader = ({ currentDate, onDateDown, onDateUp }) => {
+  return (
+    <div className="side-nav-calendar-date-header">
+      <IconButton
+        variant="white"
+        icon="chevron-left"
+        onClick={onDateDown}
+        className="chevron-btn"
+      />
+
+      <p className="h7">
+        {currMonthName(currentDate)} , {currYearName(currentDate)}
+      </p>
+
+      <IconButton
+        variant="white"
+        onClick={onDateUp}
+        icon="chevron-right"
+        className="chevron-btn"
+      />
+    </div>
+  );
+};
