+++ src/common/navigation/calendar/date-header/styles.scss
@@ -1,0 +1,16 @@
+.side-nav-calendar-date-header {
+  width: 100%;
+  height: 40px;
+  padding: 0px 4px;
+  background-color: #2b2b2c;
+
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: space-between;
+
+  .chevron-btn {
+    width: 24px;
+    height: 24px;
+  }
+}
