+++ src/common/navigation/styles.scss
@@ -53,24 +53,18 @@
     width: 100%;
     margin-top: auto;
     justify-self: flex-end;
-    // background-color: red;
+    overflow: hidden;
 
     display: flex;
     align-items: flex-start;
     flex-direction: column;
 
     .action-btn {
+      width: 248px;
       border-radius: 0px;
     }
 
-    .bg-calendar {
-      opacity: 1;
-      margin-bottom: 24px;
-      width: 100%;
-      height: 300px;
-    }
-
-    .pf-icon-btn {
+    .side-menu-icon-btn {
       width: 100%;
       height: 64px;
       padding-left: 28px;
